extract main methods from picocli #62
