+++ build.gradle.kts
@@ -14,6 +14,9 @@ allprojects {
 
 subprojects {
     group = "io.github.linguaphylo"
+    // TODO 3 versions: here, LPhyBEAST, version.xml
+    // version has to be manually adjusted to keep same between version.xml and here
+    version = "1.0.0-SNAPSHOT"
     val webSteam = "github.com/LinguaPhylo/LPhyBeast"
     val web = "https://${webSteam}"
