+++ version.xml
@@ -27,7 +27,7 @@
     <!-- Not released : contraband bdtree -->
 
     <packageapp description="A command-line program that takes an LPhy model specification including a data block, and produces a BEAST 2 XML input file."
-                class="lphybeast.LPhyBEAST"
+                class="lphybeast.LPhyBeastCMD"
                 args=""
     />
